move leave tournament check into services
