+++ app/services/games/tournament/tournamentValidation/validTournamentConnectionCheck.ts
@@ -5,7 +5,9 @@ const validTournamentConnectionCheck = (
   tournamentId: string,
 ) => {
   if (!tournaments.has(tournamentId) === false) {
-    throw new Error("Tournament does not exist");
+    throw new Error(
+      "[validTournamentConnectionCheck] Tournament does not exist",
+    );
   }
   if (tournaments.get(tournamentId)?.isMemberInTournament(memberId) === false) {
     throw new Error("You are not a member of this tournament");
